readline: move Interface to separate file
This commit moves the readline Interface class to its own
file. All other readline logic is moved into the internal
readline utils file.
