+++ node.gyp
@@ -170,6 +170,7 @@
       'lib/internal/process/report.js',
       'lib/internal/process/task_queues.js',
       'lib/internal/querystring.js',
+      'lib/internal/readline/interface.js',
       'lib/internal/readline/utils.js',
       'lib/internal/repl.js',
       'lib/internal/repl/await.js',
